move scale logic to _PolygonScaleHandler.svelte
